Works on single delete, TODO multiple deletes and add bundle name that is going to be deleted

+++ app/src/main/java/app/revanced/manager/ui/component/bundle/BundleItem.kt
@@ -7,12 +7,15 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.size
 import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.outlined.Delete
 import androidx.compose.material.icons.outlined.ErrorOutline
 import androidx.compose.material.icons.outlined.Warning
+import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Icon
 import androidx.compose.material3.ListItem
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -50,11 +53,23 @@ fun BundleItem(
     val name by bundle.nameState
 
     if (viewBundleDialogPage) {
+        var showDeleteConfirmationDialog by remember { mutableStateOf(false) }
+
+        if (showDeleteConfirmationDialog) {
+            BundleDeleteDialog(
+                onDismiss = { showDeleteConfirmationDialog = false },
+                onConfirm = {
+                    onDelete()
+                    viewBundleDialogPage = false
+                }
+            )
+        }
+
         BundleInformationDialog(
             onDismissRequest = { viewBundleDialogPage = false },
             onDeleteRequest = {
-                viewBundleDialogPage = false
-                onDelete()
+                showDeleteConfirmationDialog = true
+                //
             },
             bundle = bundle,
             onUpdate = onUpdate,
@@ -106,5 +121,33 @@ fun BundleItem(
                 version?.let { Text(text = it) }
             }
         },
+    )
+}
+
+@Composable
+private fun BundleDeleteDialog(
+    onDismiss: () -> Unit,
+    onConfirm: () -> Unit
+) {
+    AlertDialog(
+        onDismissRequest = onDismiss,
+        dismissButton = {
+            TextButton(onDismiss) {
+                Text(stringResource(R.string.cancel))
+            }
+        },
+        confirmButton = {
+            TextButton(
+                onClick = {
+                    onConfirm()
+                    onDismiss()
+                }
+            ) {
+                Text(stringResource(R.string.confirm))
+            }
+        },
+        title = { Text(stringResource(R.string.delete_bundle_dialog_title)) },
+        icon = { Icon(Icons.Outlined.Delete, null) },
+        text = { Text(stringResource(R.string.delete_bundle_dialog_description)) }
     )
 }
